adding tab and new view to project

+++ app.py
@@ -114,7 +114,7 @@ with gr.Blocks() as demo:
     def generate_mask(image, selected_points, radio_mask):
         predictor.set_image(image)
         input_points = np.array(selected_points, dtype=object)[0,:]
-        input_labels = np.array(selected_points, dtype=int)[:,1]
+        input_labels = np.array(selected_points)[:,1]
         print(selected_points)
         masks, _, _ = predictor.predict(
             point_coords=input_points,
